[Chore] #11 - 바텀 시트 관련 PR 리뷰 반영
