+++ TOASTER-iOS/Global/Components/ToasterBottomSheet/BottomType.swift
@@ -1,0 +1,30 @@
+//
+//  BottomType.swift
+//  TOASTER-iOS
+//
+//  Created by 민 on 1/3/24.
+//
+
+import UIKit
+
+enum BottomType {
+    case white, gray
+    
+    var color: UIColor {
+        switch self {
+        case .white:
+            return .toasterWhite
+        case .gray:
+            return .gray50
+        }
+    }
+    
+    var alignment: NSTextAlignment {
+        switch self {
+        case .white:
+            return .left
+        case .gray:
+            return .center
+        }
+    }
+}
